refactor(weather): split cache helpers into weather-cache module
